refactor: replace Instructure component with main component and update imports
